identity-appsupport: New Compose Multiplatform lib for apps. (#686)
The idea is that reusable composables and other UI components will go
into this library. Users of this library will be end-user apps like
`wallet` and the idea is that we'll use `testapp` to test things that
aren't easily testable through unit tests.

For now move `PassphraseEntryField` into this library. Also add a page
to `testapp` to test all main ways this composable can be used.

Test: Manually tested on Android and iOS.
Test: ./gradlew check
Test: ./gradlew connectedCheck

Signed-off-by: David Zeuthen <zeuthen@google.com>