Fix use of Semantic Versioning. (#1000)
According to Semantic Versioning 2.0 prereleases should use the next
version number which is contrary to what we're doing today. So fix
this up. Also gracefully handle the case where it's not possible to
reach apps.multipaz.org - for example if there is no Internet
connectivity - instead of crashing.

With this change version numbers will change from
e.g. 0.90-48.g57b479c to e.g. 0.91.0-pre.48.84706e70.

This introduces an additional post-release version bump step as part
of making a release.

Test: Manually tested.

Signed-off-by: David Zeuthen <zeuthen@google.com>