DigitalCredentials: Change API to take protocol and data for response. (#1136)
The response string being sent back to Credman is expected to a string
with the encoded version of the DigitalCredential object according to
W3C DC spec, for example

```
{
  protocol: "openid4vp-v1-signed"
  data: "<data>"
}
```

but we were sending back only `data`, not the enclosing object with
`protocol`. The receiving end has in Android has a workaround for this
because the spec used to require only `data`. This workaround isn't
going to stay around forever so update to what the spec calls for.

Manually tested with all W3C DC protocols we implement which is:
OpenID4VP 24 and 29, 18013-7 Annex C, Preview, ARF.

Test: Manually tested.

Signed-off-by: David Zeuthen <zeuthen@google.com>