Add support for native version of navigator.credentials.get(). (#1274)
Concretely, this includes a new `request()` method on the existing
`DigitalCredentials` type which can be used to request credentials
from local applications on the same device. Also add support for this
in samples/testapp in a new "W3C Digital Credentials Request" screen
which works a lot like https://verifier.multipaz.org except that it's
a native app requesting from local apps on the device.

Since user interaction is needed for this, introduce new `UiContext`
multiplatform class with concrete `AndroidUiContext` implementation
for Android. This is used by `DigitalCredentials.request()` to get a
reference to the Activity to bounce back to when having obtained the
response from the wallet.

Also factor out existing verification code to `VerificationUtil`
object and change multipaz-verifier-server to use this.

Test: Manually tested.

Signed-off-by: David Zeuthen <zeuthen@google.com>