js: Wrap Web Crypto and other JS Browser APIs (#1477)
This includes changing functions in `Crypto`, `Compression`, and
elsewhere to be suspend which in turn slightly changes the shape of
other APIs to also be suspend. We implemented just enough to get all
unit tests to pass and doesn't include browser-specific
implementations of `Storage` or `SecureArea`.

As part of this, implement PEM parsing and generation entirely in KMP
and remove internal platform-specific wrappers in `Crypto`. Also add
unit tests for platform-specific implementation of `X509CertChain`
signature checking.

Since we can't easily change multipaz-android-legacy to be suspend,
make copies of old non-suspend APIs from multipaz `Crypto`, `Hkdf`,
and `SessionEncryption` APIs there. Code-duplication is not optimal
but this library is going away fairly soon as the consuming downstream
is in the process of switching away from it to mainline Multipaz
libraries

Test: All tests pass.
Test: Manually tested testapp on both Android and iOS.

Signed-off-by: David Zeuthen <zeuthen@google.com>